collect convenience extensions for SampleApp in one file
